pyproject.toml
[project]
name = "reiproxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kivymd",
]

[tool.uv.sources]
kivymd = { url = "https://github.com/kivymd/KivyMD/archive/master.zip" }
